Check for None using "is" instead of equality.
My machine returns an error when e.g. an array is checked for None type using equality instead of "is". i.e. array==None returns an array of True/False values, which doesn't work in "if" statements (asks for "any" or "all"). It seems like the PEP recommendation for NoneType checking is to use "is" rather than equality. So, I changed the few cases of this. https://www.python.org/dev/peps/pep-0008/#programming-recommendations