Discard pix under mask to compute temporal eigv
Not taking into account the pixels under the mask when computing temporal pc's required to discard those pixels from the vectorized images.

This is done in prepare_matrix() :

If, discard_mask_pix is set to True,
Then, mask_circle is asked to return a boolean mask instead of a masked array
The array is truncated accordingly to the boolean mask before being flattened.

Then, _svd_wrapper takes into account the spatially truncated vectors to compute the full temporal pc's. Then project the full matrix (with regular 0-filled mask) on the pc's :+1: