Replacement of sci_cube_skysub with DIKL
Replace the content of sci_cube_skysub with the DIKL procedure as explained in [REN23].
In pca_fullfr, the mask_rdi has been added to _adi_rdi_pca and is used along with sci_cube_skysub when it is provided. If not, it follows the procedure in place.
The argument mask_rdi in docstring has been updated accordingly.