Torch FFT Rotation & Juillard23 IPCA
1-Adding Torch for Faster FFT Rotation Mode
Fft rotation using vip-fft mode use numpy library. The use of Torch instead of Numpy consistently improves speed. New imlib='torch-fft'.

2-Adding “Juillard23” Mode to IPCA
The “Juillard23” mode has been added to the IPCA mode, enabling users to utilize the implementation from Juillard et al. (2023/2024). This method has no additional options for significant signal extraction and works exclusively with Torch, making it faster but also more prone to propagating noise and disk flux. The mode can return a stim map and residual map if full-output is set to true. This option is only compatible with 'ADI' and 'ARDI'; an error is raised if the options do not match. Other parameters, such as imbil and interpolation, are ignored. This option only uses nit, ncomp, pup_mask_center_px, cube_ref, cube, and angle_list. Installation of the GreeDS package (via pip install GreeDS) is required for this option. In particular, the threshold is always set to 0, and STIM is only computed at the end and is not part of the iterative process.